LPD-8519 Merge SharingNotificationHelper with NotificationsSharingEntryLocalServiceWrapper and apply to usages. Keep SharingEntryLocalService unused reference, the reference is necessary to the ServiceWrapperServiceTrackerCustomizer into ServiceWrapperRegistry
